update/ adjust component in GroupInfomation
